switched a bunch of "var" to "let" + added object length check to "isEmpty()" + renamed "networking.js" to "net.js"
