+++ src/files.js
@@ -1,6 +1,6 @@
 function logger(path, content, options) {
-    var fs = require("fs");
-    var isEmpty = require("./misc").isEmpty;
+    let fs = require("fs");
+    let isEmpty = require("./misc").isEmpty;
 
     if(isEmpty(path) || isEmpty(content) || typeof path != "string" || typeof content != "string"){console.log("path and/or content are empty or of the wrong type");return;}
 
@@ -20,13 +20,13 @@ module.exports.logger = logger;
 
 
 function readdirRecursive(folder, callback) { // modified version of https://stackoverflow.com/a/5827895/8602926
-    var fs = require("fs");
-    var path = require("path");
-    var walk = function(dir, done) {
-        var results = [];
+    let fs = require("fs");
+    let path = require("path");
+    let walk = function(dir, done) {
+        let results = [];
         fs.readdir(dir, function(err, list) {
             if(err) return done(err);
-            var pending = list.length;
+            let pending = list.length;
             if(!pending) return done(null, results);
             list.forEach(function(file) {
                 file = path.resolve(dir, file);
@@ -50,20 +50,16 @@ module.exports.readdirRecursive = readdirRecursive;
 
 
 function readdirRecursiveSync(folder) { // from https://stackoverflow.com/a/16684530/8602926
-    var fs = require("fs");
-    var walk = function(dir) {
-        var results = [];
-        var list = fs.readdirSync(dir);
+    let fs = require("fs");
+    let walk = function(dir) {
+        let results = [];
+        let list = fs.readdirSync(dir);
         list.forEach(function(file) {
             file = dir + '/' + file;
-            var stat = fs.statSync(file);
-            if (stat && stat.isDirectory()) { 
-                /* Recurse into a subdirectory */
+            let stat = fs.statSync(file);
+            if (stat && stat.isDirectory())
                 results = results.concat(walk(file));
-            } else { 
-                /* Is a file */
-                results.push(file);
-            }
+            else results.push(file);
         });
         return results;
     }
