+++ src/rng.js
@@ -122,12 +122,12 @@ const randRange = (min, max) => {
 const generateUUID = (uuid_format) => {
     if(isEmpty(uuid_format) || typeof uuid_format != "string") return "wrong attribute type - has to be of type string";
 
-    var d = new Date().getTime();
+    let d = new Date().getTime();
     if (typeof performance !== 'undefined' && typeof performance.now === 'function'){
         d += performance.now();
     }
     return uuid_format.replace(/[xy]/g, function (c) {
-        var r = (d + Math.random() * 16) % 16 | 0;
+        let r = (d + Math.random() * 16) % 16 | 0;
         d = Math.floor(d / 16);
         return (c === 'x' ? r : (r & 0x3 | 0x8)).toString(16);
     });
